+++ Models/ExpensesModel.cs
@@ -5,7 +5,7 @@ public class ExpensesModel
     // Variables
     public Guid Id { get; init; }
     public string NameExpense { get; private set; }
-    public int AmountExpense { get; private set; }
+    public decimal AmountExpense { get; private set; }
     public string? DescriptionExpense { get; private set; }
     public string? CategoryExpense { get; private set; }
     public DateTime DateExpense { get; private set; }
@@ -13,7 +13,7 @@ public class ExpensesModel
     // Construtor
     
     public ExpensesModel() { }
-    public ExpensesModel(string nameExpense, int amountExpense, string descriptionExpense, string categoryExpense)
+    public ExpensesModel(string nameExpense, decimal amountExpense, string descriptionExpense, string categoryExpense)
     {
         Id = Guid.NewGuid();
         NameExpense = nameExpense;
@@ -23,7 +23,7 @@ public class ExpensesModel
         DateExpense = DateTime.UtcNow;
     }
 
-    public void ChangeValues(string nameExpense, int amountExpense, string descriptionExpense, string categoryExpense)
+    public void ChangeValues(string nameExpense, decimal amountExpense, string descriptionExpense, string categoryExpense)
     {
         NameExpense = nameExpense;
         AmountExpense = amountExpense;
